move example file
